requirements.txt
pyserial
matplotlib
pandas
keyboard
crc
pynput
plotly
magpylib
numpy-stl
pyyaml
wandb
scipy
